+++ src/main/scala/com/scylladb/migrator/config/CopyType.scala
@@ -1,0 +1,7 @@
+package com.scylladb.migrator.config
+
+sealed trait CopyType
+object CopyType {
+  case object WithTimestampPreservation extends CopyType
+  case object NoTimestampPreservation extends CopyType
+}
