+++ src/main/scala/com/scylladb/migrator/config/SourceSettings.scala
@@ -1,0 +1,65 @@
+package com.scylladb.migrator.config
+
+import cats.implicits._
+import io.circe.syntax._
+import io.circe.{ Decoder, DecodingFailure, Encoder, Json, ObjectEncoder }
+
+sealed trait SourceSettings
+object SourceSettings {
+  case class Cassandra(host: String,
+                       port: Int,
+                       credentials: Option[Credentials],
+                       keyspace: String,
+                       table: String,
+                       splitCount: Option[Int],
+                       connections: Option[Int],
+                       fetchSize: Int)
+      extends SourceSettings
+  case class Parquet(path: String, credentials: Option[AWSCredentials]) extends SourceSettings
+
+  implicit val decoder: Decoder[SourceSettings] = Decoder.instance { cursor =>
+    cursor.get[String]("type").flatMap {
+      case "cassandra" | "scylla" =>
+        Decoder
+          .forProduct8(
+            "host",
+            "port",
+            "credentials",
+            "keyspace",
+            "table",
+            "splitCount",
+            "connections",
+            "fetchSize")(Cassandra.apply)
+          .apply(cursor)
+      case "parquet" =>
+        Decoder
+          .forProduct2("path", "credentials")(Parquet.apply)
+          .apply(cursor)
+      case otherwise =>
+        Left(DecodingFailure(s"Unknown source type: ${otherwise}", cursor.history))
+    }
+  }
+
+  implicit val encoder: Encoder[SourceSettings] = Encoder.instance {
+    case s: Cassandra =>
+      Encoder
+        .forProduct8(
+          "host",
+          "port",
+          "credentials",
+          "keyspace",
+          "table",
+          "splitCount",
+          "connections",
+          "fetchSize")(Cassandra.unapply(_: Cassandra).get)
+        .encodeObject(s)
+        .add("type", Json.fromString("cassandra"))
+        .asJson
+    case s: Parquet =>
+      Encoder
+        .forProduct2("path", "credentials")(Parquet.unapply(_: Parquet).get)
+        .encodeObject(s)
+        .add("type", Json.fromString("parquet"))
+        .asJson
+  }
+}
